% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-tidyeval.R
\name{compat_as_lazy}
\alias{compat_as_lazy}
\alias{compat_as_lazy_dots}
\title{Convert tidyeval-compatible inputs to lazy objects}
\usage{
compat_as_lazy(quo)

compat_as_lazy_dots(...)
}
\arguments{
\item{quo}{A quosure.}

\item{...}{Quasiquoted inputs.}
}
\description{
Convert tidyeval-compatible inputs to lazy objects
}
