% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-tidyeval.R
\name{compat_lazy}
\alias{compat_lazy}
\alias{compat_lazy_dots}
\title{Convert lazyeval-compatible inputs to quosures}
\usage{
compat_lazy(lazy, env = caller_env(), warn = TRUE)

compat_lazy_dots(dots, env, ..., .named = FALSE)
}
\arguments{
\item{lazy, dots, ...}{Lazyeval-compatible inputs, e.g. formulas,
strings, lazy objects, etc.}

\item{env}{A default environment for the returned quosure.}

\item{warn}{Whether to issue a deprecation warning.}

\item{.named}{Whether to give default names to unnamed inputs.}
}
\description{
Convert lazyeval-compatible inputs to quosures
}
