% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{f_from_lhs}
\alias{f_from_lhs}
\alias{f_from_rhs}
\title{Extract parts of a two-sided formula}
\usage{
f_from_lhs(f)

f_from_rhs(f)
}
\arguments{
\item{f}{A two-sided formula}
}
\value{
A one-sided formula
}
\description{
These functions extract a one-sided formula from a two-sided one.
}

